chore(app): Move agreement list components
